requirements.txt
openai
chromadb
rank_bm25
sentence-transformers
networkx
python-dotenv
colorama
tf-keras